# [PERF GAIN?!!] Make document / value usable multiple times