This makes document and value more useful, allowing you to first try to parse a value as an array, and then as an object.

## Performance

I absolutely cannot say *why*, but this version actually fulfills what I consider to be the true promise of ondemand: the reduction of branches. It brings the number of branches down from 3500 to 500, increasing speed to 6GB/s!!!! (It's still checking the results against the DOM, so as long as that check is still working, it may still be accurate.)

@lemire If you would be so kind, I'd love some confirmation / checks on this result. I'm having trouble believing it, as much as I want to. At the least, I would expect the branch predictor to suck at the string or number parsing loops. Could the branches really have been somewhere else?

### Haswell gcc10.2 (Skylake)

Compiled with `cmake SIMDJSON_IMPLEMENTATION=haswell CMAKE_CXX_FLAGS="-march=native" ..`

```c++
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations best_branch_miss best_bytes_per_sec best_cache_miss best_cache_ref best_cycles best_cycles_per_byte best_docs_per_sec best_frequency best_instructions best_instructions_per_byte best_instructions_per_cycle best_items_per_sec branch_miss      bytes bytes_per_second cache_miss  cache_ref     cycles cycles_per_byte docs_per_sec  frequency instructions instructions_per_byte instructions_per_cycle      items items_per_second
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
PartialTweets<OnDemand>      110985 ns       110984 ns         6294              538           6.13556G               0         26.41k    380.178k             0.602009          9.71562k       3.69367G          1.38172M                    2.18794                      3.6344                  0     568.854   631.515k       5.29934G/s  0.0255799   26.5123k   381.003k        0.603317   9.01028k/s 3.43295G/s     1.38172M               2.18794                3.62653          0              0/s [best: throughput=  6.14 GB/s doc_throughput=  9715 docs/s instructions=     1381719 cycles=      380178 branch_miss=     538 cache_miss=       0 cache_ref=     26410 items=         0 avg_time=    103218 ns]
PartialTweets<Iter>          389673 ns       389671 ns         1794           3.046k           1.66902G               0        54.857k    1.39693M              2.21203          2.64289k       3.69193G          4.41354M                    6.98882                     3.15946           264.289k    3.22385k   631.515k       1.50933G/s   0.269788   54.9803k   1.40863M         2.23056   2.56627k/s 3.61493G/s     4.41354M               6.98882                3.13321        100       256.627k/s [best: throughput=  1.67 GB/s doc_throughput=  2642 docs/s instructions=     4413544 cycles=     1396930 branch_miss=    3046 cache_miss=       0 cache_ref=     54857 items=       100 avg_time=    381762 ns]
PartialTweets<Dom>           270718 ns       270718 ns         2584           3.618k           2.41985G               0        87.043k    963.578k              1.52582          3.83181k       3.69225G          2.91909M                    4.62236                     3.02943           383.181k    3.71913k   631.515k       2.17253G/s  0.0731424   87.1628k    969.03k         1.53445   3.69387k/s 3.57948G/s     2.91909M               4.62236                3.01238        100       369.387k/s [best: throughput=  2.42 GB/s doc_throughput=  3831 docs/s instructions=     2919092 cycles=      963578 branch_miss=    3618 cache_miss=       0 cache_ref=     87043 items=       100 avg_time=    262593 ns]
Creating a source file spanning 44921 KB 
LargeRandom<Dom>           87772101 ns     87771741 ns            8         908.793k           524.903M        11.0279M       15.2321M    323.242M               7.0272           11.4112        3.6886G          970.316M                    21.0944                     3.00182           11.4112M    908.714k   45.9988M       499.795M/s   11.0372M   15.2312M   323.728M         7.03774    11.3932/s 3.68829G/s     970.316M               21.0944                2.99732      1000k       11.3932M/s [best: throughput=  0.52 GB/s doc_throughput=    11 docs/s instructions=   970315577 cycles=   323242407 branch_miss=  908793 cache_miss=11027918 cache_ref=  15232088 items=   1000000 avg_time=  87758665 ns]
LargeRandomSum<Dom>        87310215 ns     87309740 ns            8         908.769k            527.24M        10.3713M       14.3767M    321.855M              6.99702           11.4621       3.68911G          974.316M                    21.1813                     3.02719           11.4621M    909.225k   45.9988M       502.439M/s   10.4238M   14.3788M   322.043M         7.00113    11.4535/s 3.68851G/s     974.316M               21.1813                3.02542      1000k       11.4535M/s [best: throughput=  0.53 GB/s doc_throughput=    11 docs/s instructions=   974315577 cycles=   321854511 branch_miss=  908769 cache_miss=10371311 cache_ref=  14376702 items=   1000000 avg_time=  87297241 ns]
LargeRandom<OnDemand>      13336425 ns     13336558 ns           52         283.171k           3.45602G        2.51136M       3.54749M    49.1038M               1.0675           75.1329       3.68931G          144.703M                     3.1458                     2.94688                  0    283.737k   45.9988M        3.2122G/s   2.52584M   3.54784M   49.1557M         1.06863    74.9819/s 3.68578G/s     144.703M                3.1458                2.94376          0              0/s [best: throughput=  3.46 GB/s doc_throughput=    75 docs/s instructions=   144702733 cycles=    49103777 branch_miss=  283171 cache_miss= 2511357 cache_ref=   3547489 items=         0 avg_time=  13325088 ns]
LargeRandomSum<OnDemand>   13329033 ns     13328388 ns           52         282.575k           3.45963G        2.50835M       3.54703M     49.052M              1.06638           75.2113       3.68927G          144.703M                     3.1458                     2.94999                  0    283.593k   45.9988M       3.21417G/s   2.52355M   3.54794M   49.1235M         1.06793    75.0278/s 3.68563G/s     144.703M                3.1458                2.94569          0              0/s [best: throughput=  3.46 GB/s doc_throughput=    75 docs/s instructions=   144702729 cycles=    49052009 branch_miss=  282575 cache_miss= 2508346 cache_ref=   3547030 items=         0 avg_time=  13317900 ns]
LargeRandom<Iter>          54251484 ns     54251078 ns           13         880.429k           850.609M        5.59487M       7.89165M    199.499M              4.33706            18.492       3.68914G          570.695M                    12.4067                     2.86063            18.492M    879.324k   45.9988M       808.608M/s   5.66611M   7.89564M   200.095M            4.35    18.4328/s 3.68831G/s     570.695M               12.4067                2.85212      1000k       18.4328M/s [best: throughput=  0.85 GB/s doc_throughput=    18 docs/s instructions=   570694598 cycles=   199499403 branch_miss=  880429 cache_miss= 5594870 cache_ref=   7891653 items=   1000000 avg_time=  54240015 ns]
LargeRandomSum<Iter>       55490086 ns     55489847 ns           13         883.247k           832.238M        5.13275M       7.05436M    203.882M              4.43234           18.0926       3.68876G          581.696M                    12.6459                      2.8531           18.0926M    881.331k   45.9988M       790.556M/s   5.08156M   7.04886M   204.656M         4.44916    18.0213/s 3.68817G/s     581.696M               12.6459                2.84231      1000k       18.0213M/s [best: throughput=  0.83 GB/s doc_throughput=    18 docs/s instructions=   581696111 cycles=   203882369 branch_miss=  883247 cache_miss= 5132749 cache_ref=   7054356 items=   1000000 avg_time=  55478018 ns]
Creating a source file spanning 134087 KB 
Kostya<Dom>                85336266 ns     85337119 ns            8         456.934k           1.61094G        15.7915M       22.1309M    314.421M              2.28995           11.7325       3.68896G          936.468M                    6.82035                     2.97838           6.15123M    453.718k   137.305M       1.49847G/s   15.8495M   22.1507M   314.739M         2.29226    11.7182/s 3.68819G/s     936.468M               6.82035                2.97538   524.288k       6.14373M/s [best: throughput=  1.61 GB/s doc_throughput=    11 docs/s instructions=   936467834 cycles=   314421352 branch_miss=  456934 cache_miss=15791535 cache_ref=  22130927 items=    524288 avg_time=  85323257 ns]
KostyaSum<Dom>             84917472 ns     84916847 ns            8          452.23k           1.62079G        15.5299M       21.7797M    312.525M              2.27614           11.8043       3.68916G          938.565M                    6.83562                     3.00316           6.18887M    455.497k   137.305M       1.50589G/s    15.579M   21.7169M   313.193M           2.281    11.7762/s 3.68823G/s     938.565M               6.83562                2.99677   524.288k       6.17413M/s [best: throughput=  1.62 GB/s doc_throughput=    11 docs/s instructions=   938564987 cycles=   312525386 branch_miss=  452230 cache_miss=15529878 cache_ref=  21779734 items=    524288 avg_time=  84903775 ns]
Kostya<OnDemand>           25374606 ns     25374850 ns           28         155.387k           5.41744G        4.93244M       6.78752M     93.502M              0.68098           39.4555       3.68917G          290.217M                    2.11367                     3.10386                  0    155.442k   137.305M       5.03945G/s   4.94441M   6.78768M   93.5632M        0.681426    39.4091/s 3.68724G/s     290.217M               2.11367                3.10183          0              0/s [best: throughput=  5.42 GB/s doc_throughput=    39 docs/s instructions=   290216998 cycles=    93501975 branch_miss=  155387 cache_miss= 4932442 cache_ref=   6787518 items=         0 avg_time=  25363465 ns]
KostyaSum<OnDemand>        25373695 ns     25373474 ns           28         155.303k           5.41688G        4.93959M       6.78811M    93.5123M             0.681055           39.4514        3.6892G          290.217M                    2.11367                     3.10352                  0      155.4k   137.305M       5.03972G/s   4.95093M   6.78829M   93.5588M        0.681394    39.4112/s 3.68727G/s     290.217M               2.11367                3.10197          0              0/s [best: throughput=  5.42 GB/s doc_throughput=    39 docs/s instructions=   290217001 cycles=    93512314 branch_miss=  155303 cache_miss= 4939587 cache_ref=   6788108 items=         0 avg_time=  25362328 ns]
Kostya<Iter>               59900132 ns     59899549 ns           12         457.862k            2.2957G        10.1502M       13.9903M    220.639M              1.60693           16.7197       3.68903G          642.015M                    4.67583                      2.9098           8.76596M    458.064k   137.305M       2.13483G/s   10.1909M   13.9473M   220.921M         1.60898    16.6946/s 3.68819G/s     642.015M               4.67583                2.90609   524.288k       8.75279M/s [best: throughput=  2.30 GB/s doc_throughput=    16 docs/s instructions=   642015217 cycles=   220638961 branch_miss=  457862 cache_miss=10150222 cache_ref=  13990264 items=    524288 avg_time=  59888148 ns]
KostyaSum<Iter>           117866389 ns    117867545 ns            6         458.421k           1.16793G        9.96235M       13.6368M    433.705M               3.1587           8.50609       3.68913G           1.3173G                    9.59395                     3.03731           4.45964M    458.104k   137.305M       1110.94M/s   9.98615M   13.6385M   434.782M         3.16654     8.4841/s 3.68873G/s      1.3173G               9.59395                3.02979   524.288k       4.44811M/s [best: throughput=  1.17 GB/s doc_throughput=     8 docs/s instructions=  1317297836 cycles=   433704863 branch_miss=  458421 cache_miss= 9962348 cache_ref=  13636832 items=    524288 avg_time= 117853309 ns]
```